Update perf tests
